Extract stream and URL logic from core API/Endpoint
The stream and url_for methods are not core methods, so we've extracted
them into their own modules.

Stream:

  class StreamAPI < Crepe::API
    extend Crepe::Streaming

    stream do
      # Streaming!
    end
  end

URLFor:

  class ReferentialAPI < Crepe::API
    helper Crepe::Helper::URLFor

    get do
      { _links: { users: url_for(:users) } }
    end
  end

This is also a first push into requiring Ruby 2.0. We're using `prepend`
to simplify the way stream rendering is mixed in (and adding the ability
to prepend helpers with `helper HelperModule, prepend: true`), and we're
using `**named_arguments` to simplify the `url_for` logic.
