+++ lib/crepe/helper.rb
@@ -1,0 +1,7 @@
+module Crepe
+  module Helper
+
+    autoload :URLFor, 'crepe/helper/url_for'
+
+  end
+end
